requirements.txt
requests
pyyaml
torch
opencv-python
pandas
numpy
ultralytics
segment_anything
torchvision
scikit-image
folium
geopandas
KeplerGl
osmnx